ECF_4

+++ ECF_4_Spring/src/main/java/com/example/ECF_4_Spring/controller/JobOfferController.java
@@ -33,6 +33,10 @@ public class JobOfferController {
     @GetMapping("/joboffer")
     public String home(Model model){
         System.out.println(httpSession.getAttribute("user"));
+        if((httpSession.getAttribute("user")) == null || httpSession.getAttribute("user").equals("null")){
+            System.out.println("toto");
+            return "redirect:/login";
+        };
         User toto = userService.findByMailAndPassword(((User) httpSession.getAttribute("user")).getEmail(), ((User) httpSession.getAttribute("user")).getPassword());
         model.addAttribute("joboffers", jobOfferService.findAll());
         model.addAttribute("user", toto);
@@ -40,12 +44,20 @@ public class JobOfferController {
     }
     @GetMapping("/joboffer/add")
     public String addForm(Model model){
+        if((httpSession.getAttribute("user")) == null){
+            System.out.println("toto");
+            return "redirect:/";
+        }
         model.addAttribute("joboffer", new JobOffer());
         return "jobOfferForm";
     }
 
     @PostMapping("/joboffer/add")
     public String addPerson(@ModelAttribute("joboffer") JobOffer jobOffer){
+        if((httpSession.getAttribute("user")) == null){
+            System.out.println("toto");
+            return "redirect:/";
+        }
 
         if (jobOffer.getId() == null)  {
            jobOfferService.save(jobOffer);
@@ -56,17 +68,29 @@ public class JobOfferController {
     }
     @GetMapping("/joboffer/delete/{idoffer}")
     public String delete(@PathVariable("idoffer") Long idoffer){
+        if((httpSession.getAttribute("user")) == null){
+            System.out.println("toto");
+            return "redirect:/";
+        };
         jobOfferService.delete(jobOfferService.findById(idoffer));
         return "redirect:/joboffer";
     }
     @GetMapping("/joboffer/update/{idoffer}")
     public String formUpdate(@PathVariable("idoffer") Long idoffer, Model model){
+        if((httpSession.getAttribute("user")) == null){
+            System.out.println("toto");
+            return "redirect:/";
+        };
         JobOffer jobOffer = jobOfferService.findById(idoffer);
         model.addAttribute("joboffer", jobOffer);
         return "jobOfferForm";
     }
     @GetMapping("/joboffer/postuler/{idjob}")
     public String postuler(@PathVariable("idjob") Long idjob, Model model){
+        if((httpSession.getAttribute("user")) == null){
+            System.out.println("toto");
+            return "redirect:/";
+        };
         JobOffer jobOffer = jobOfferService.findById(idjob);
         model.addAttribute("joboffer", jobOffer);
         model.addAttribute("appointements", new Appointement());
@@ -74,6 +98,10 @@ public class JobOfferController {
     }
     @PostMapping("/joboffer/postuler/{idjob}")
     public String postuler(@PathVariable("idjob") Long idjob, Appointement appointement){
+        if((httpSession.getAttribute("user")) == null){
+            System.out.println("toto");
+            return "redirect:/joboffer";
+        };
         System.out.println(idjob);
         JobOffer jobOffer = jobOfferService.findById(idjob);
 
@@ -93,14 +121,14 @@ public class JobOfferController {
 
     @GetMapping("/joboffer/appointenement")
     public String appointement(Model model){
+        if((httpSession.getAttribute("user")) == null){
+            System.out.println("toto");
+            return "redirect:/";
+        };
         User user = userService.findByMailAndPassword(((User) httpSession.getAttribute("user")).getEmail(), ((User) httpSession.getAttribute("user")).getPassword());
         model.addAttribute("user",user);
         model.addAttribute("appointements", appointementService.findAll());
         return "listappointement";
     }
-    @GetMapping("/joboffer/appointenement/delete/{idrdv}")
-    public String deleteRdv(@PathVariable("idrdv") Long idrdv){
-        appointementService.delete(appointementService.findById(idrdv));
-        return "redirect:/joboffer";
-    }
+
 }
